Move waveform UI to design system library (#1649)
