+++ libraries/designsystem/src/main/kotlin/io/element/android/libraries/designsystem/components/media/WaveformPlaybackView.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.features.messages.impl.voicemessages.timeline
+package io.element.android.libraries.designsystem.components.media
 
 import android.view.MotionEvent
 import androidx.compose.animation.core.animateFloatAsState
