+++ tests/uitests/src/test/snapshots/images/ui_S_t[l.designsystem.components.media_null_WaveformPlaybackView-D_0_null,NEXUS_5,1.0,en].png
